refactor: 관심사 분리 #8
