+++ src/domain/review/repo/short-review.repository.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@nestjs/common';
 import { Prisma, PrismaClient } from '@prisma/client';
-import { ShortReviewResponseDto } from '../dto/review-response.dto';
 import { IShortReview } from '../reviews.interface';
 
 @Injectable()
@@ -11,7 +10,7 @@ export class ShortReviewRepository extends PrismaClient {
     pageSize?: number,
     sortKey?: string,
     sortDir?: string,
-  ): Promise<ShortReviewResponseDto[]> {
+  ): Promise<IShortReview[]> {
     const prismaQuery: {
       where: Prisma.ReviewWhereInput;
       skip: number;
@@ -28,7 +27,7 @@ export class ShortReviewRepository extends PrismaClient {
       },
     };
 
-    const shortReviews: IShortReview[] = await this.review.findMany({
+    return this.review.findMany({
       select: {
         id: true,
         memberId: true,
@@ -44,13 +43,5 @@ export class ShortReviewRepository extends PrismaClient {
       },
       ...prismaQuery,
     });
-
-    const shortReviewDtoList: ShortReviewResponseDto[] = [];
-    shortReviews.forEach((shortReview) => {
-      const shortReviewResponseDto = new ShortReviewResponseDto(shortReview);
-      shortReviewDtoList.push(shortReviewResponseDto);
-    });
-
-    return shortReviewDtoList;
   }
 }
